Process quest loot separate from normal loot
This is a potential fix for #1635. It ensures that quest loot gets
processed, even if the number of potential drops is 0. This also
mitigates any problems that may have occurred by mixing the quest loot
into the normal loot table.

Also in this commit, some of the duplicate code in checkLoot() has been
pulled out into a separate checkLootComponent() function.