Mouse move: Add 0.5s pause when transitioning from attack to movement
Since the lockAttack flag was broken, it has been removed and replaced
with this cooldown. Instead of locking the player in the attack state,
we prevent the player from moving for a brief period. The player can
escape the cooldown by releasing the MAIN1 (left mouse) button. This
allows players to adjust their aim without releasing the attack button.

Also, this commit fixes #1685.