Change behavior of "Always show loot tooltips"
When true, all tooltips are always shown, as was the case in v1.08.
When false, we use the new dynamic hiding that hides tooltips that would
be obscuring entities.

This removes the the ability to always hide all loot tooltips, but I am
working on a solution to allow toggling between all three options.