Move more interface options to the actionbar/inventory menus
The loot tooltips button was too big for how little importance (compared
to the main menu buttons) it had. So I've crammed 4 buttons into the
space where the loot tooltips button was. These control the following
options:

- Loot tooltip visibility
- Minimap display mode (new!)
- Always showing statbar labels
- Showing combat text

In addition, the auto-equip items option has been added as a "checkbox"
to the Inventory menu. It uses the image
"images/menus/buttons/auto_equip.png". The WidgetCheckBox class has been
updated to not trigger a critical fail if a custom image is missing when
the default image is available. This change will likely be applied to
the other widgets soon.

Anyway, the minimap mode setting is new. It currently has two options:
visible and hidden. However, I plan to try adding at least one zoom
level. That's why it's not a boolean value.