Use camera for mouse aim, not player position
When using mouse move, the player will now face the appropriate
direction, even if the camera is greatly offset. This is very noticable
when a slow camera speed (which is actually a high "camera_speed" value
in engine/misc.txt). The same applies for aiming of attacks.

Since these mouse actions are now relative to the camera, they can
unexpectedly create some unwanted behavior as the camera is changing. To
get around this, two new features have been added:

- Intrapmap teleports now lock the player's movement until the camera
catches up
- A (roughly) 100ms cooldown has been added to changing the player's
direction. This reduces the amount of jitter that can happen if the
camera is moving in such a way that the player quickly alternates
between two directions. Releasing the move button bypasses this
cooldown, but it shouldn't matter since the cooldown is so short to
begin with.