Refactored 'requires_hpmp_state' for Powers to check against both stats at once
An example use case for this is a restoration potion that restores both
HP and MP. In such a scenario, we only want to lock the power when the
player has both full HP and full MP. To do so, we'd use:

requires_hpmp_state=any,not_percent,100,not_percent,100

In English, the above can be read as: "If the HP is not 100% or the MP
is not 100%, allow using this power."

While removed from the documentation, passing 'hp' or 'mp' as the first
parameter still works. This is for backwards compatibility, but will
likely be deprecated in future versions.