Improved performance when checking map event hotspots
While playing Ghostlore, a map-sized event with a hotspot was slowing
the game to a crawl. The event in question was not even active due to
campaign status. So this commit fixes things by checking some things
BEFORE iterating over each tile in the hotspot.

In reality, this event likely shouldn't be using a hotspot at all, but
rather be of the 'static' activation type instead of 'on_trigger'.