Removed hard-coded restrictions on using items from the inventory
Previously, only items with the type "consumable" could be used by
right-clicking on them in the inventory. And even then, quest items were
explicitly not allowed.

Now, any item that has a power and does not match an equipment slot can
be used by right-clicking on it in the inventory.