Added movement input type dialog box
This is only displayed for new players (a flag is set in settings.txt
once a choice is made). This provides a simpler way of selecting input
method when compared to the Configuration menu. Perhaps it would be
worth adding a button to the Configuration menu to show this dialog...