Add "safe video" feature
This should help players that can not start Flare at all with the
default settings. Along with the command line option, a button has been
added to the lock file dialog. After a crash, the lock file is almost
always present, so we can use it to "detect" the crash and provide the
fallback option.