StatBlock: Optimize per-primary stat calculation in calcBase()
These changes were tested prior to the previous commit on a Ghostlore
map with lots of Power events. All the StatBlocks were doing their
logic(), which slowed down to less than 1 FPS on my laptop. After these
changes, I was able to get the framerate up to 8-12 FPS. Not great, but
a big enough improvement to warrant this change.

For reference, the change to the map StatBlocks in the previous commit
brought my FPS in the test scene up to 20-22 FPS.