Fix issues with item comparison tooltips
- Utilize tooltip caching by using a fixed array of persistent
  WidgetTooltip objects instead of creating them dynamically. This fixes
  the high CPU usage.
- The tooltips for equipped items are marked as children of target tip.
  This lets us position the children relative to their parent,
  eliminating the overlap bug.
- MenuManager::pushMatchingItemsOf() has been optimized a little. The
  calls to checkTooltip() have been removed. Instead, we only check the
  emptiness of the item_type variable of the hovered ItemStack's item.
- Fixed the spelling of "Equipped" and made it able to be translated.
- Comparison tooltips now function when not using a mouse (i.e. gamepad)

NOTE: At low resolutions, it may occur that one of the child tooltips
doesn't fix on the screen. At the moment, we simply render them anyway,
letting them get cut off. I tried hiding the child tooltip in this case,
as well as shifting the other tooltips to try fitting it in. The former
felt inconsistent, and the latter was awkward when a tooltip covered the
icon because it was too tall. I guess it can stay like this for now...