AnimationSet: Don't try to load if filename is blank
FileParser's open() uses ModManager::list(), which will get all the
files in a directory if not given a specific file. Passing an empty
string will get all the files in the mod's root! This is usually just
the settings.txt file, but that is most definitely not the behavior we
want here!