EntityBehavior: Improved handling of entering/leaving combat state
This fixes a bug where enemies would exit combat if they couldn't find
an ally character to target, even if the player was a valid target. As
a result, the change in the previous commit is no longer needed because
the enemy stays in combat when they're supposed to.

This also fixes a bug where allies that lost their target when in combat
would try to attack the player instead.