Added additional requirement properties to enemy and NPC map objects
The "invincible" properties for map enemies are still only able to check
campaign statues. I don't see the need to implement the other properties
here. I think it would just bloat this part of the mod API. However, the
code changes in this commit should make it easier to implement them
later if desired.