New method of mouse-less action bar interaction
Instead of having 4 dedicated keys for the action bar, there are now
2 keys. One toggles the focus on the action bar. The other is a generic
activation key that will eventually work in other menus as well. So
really, there is only one key dedicated to the action bar.

Navigating the action bar is now done with the standard movement keys.
Picking up an action bar item to rearrange it is done with the Accept
key.

A few other bindings have been made in InputState, but these are for
future use.