Fix minimap not updating enough when using fog-of-war
Unlike ed6c077e, this fix does not require fog tilesets to be loaded. As
it turned out, the true cause of the original bug was that the clip size
for updating the minimap texture simply wasn't big enough. As a fix, we
now use the whole visible portion of the minimap, which should still be
performant (flare-game uses 128x128).