Minimap: Use update functions for prerender
bb59a0ee wasn't a correct fix, and I was still seeing cases where the
tiles rendered in the update function would diverge from the prerendered
tiles. By using the update functions for all rendering, we get
consistent output between the prerender and the updates. And it
consolidates what was effectively duplicate code.

- Fixed center point being off by 1 pixel
- Fixed prerender when fog-of-war is disabled