Improved text rendering by blending text with a copy of itself
I was looking at some old screenshots (around v0.16) and noticed that
text rendering has decreased in quality. We were only using
TTF_RenderTextSolid() back then, but the current iteration of that
function causes aliased output. But using the "blended" version of that
function causes text to look "thin" in many places.

The only solution I found that looked like the old style was to render
the same thing twice, effectively blending with itself. The end result
is much clearer, especially for glyphs that have lots of small details.
In fact, I would say this closes #1536.