NSIS: Use "Program Files" by default instead of "Program Files (x86)"
This also fixes the UninstallPrevious function. The use of ExecWait with
the uninstall program didn't actually wait until the process was
finished (even when trying the _? flag). This could result in the
uninstaller undoing actions done by the installer itself. So instead,
there's now a function within the installer that will remove old
installations.