Added 'mouse_move_deadzone' property to engine/misc.txt
This is done to help alleviate erratic mouse-move behavior when the
cursor is near the player. As mentioned in the docs here, the deadzone
while moving should be smaller than the deadzone when stationary.

The decision to make this an engine setting was made because the values
here would change based on a mod's relationship between the player size
and the tile size. The defaults here work under the assumption that the
player occupies one whole tile.

Example usage:
mouse_move_deadzone=0.25,0.75