EventManager: Refactor EventComponent; Support floats for some types
The following event component types now support floats:
- power_damage
- chance_exec

In summary, this commit replaces the previous integers (x,y,z,a,b,c) and
float (f) with a single array of unions. The LOOT type requires the most
amount of this data with 4 ints and 1 float, so the array size has been
set to 5. As it turned out, the 'c' int wasn't being used at all, and
the 'z' int was never used with all 'x', 'y', 'a', and 'b' at the same
time. Using a union also means we can create a field for bool and avoid
having to cast ints.