Added 'resource_round_method' property to engine/combat.txt
This gives us more granular control over how the math that deals with HP
and MP is handled. Specifically, this affects cases when combat text
gets displayed, such as taking damage or healing.

The default behavior is to round to the nearest integer. This is in
contrast to the old behavior where we essentially did a 'floor' on the
value (any float calculation results were recast as int). Always
rounding down could be perceived as a bug, which is why I'm fine with
changing the default behavior here.