Add 'loot_drops_max' property to Items
This can be used to limit the number of drops in a single loot event.
This does *not* handle getting more than one copy of an item in a single
playthrough, as mentioned in issue #1672. To me, that is acceptable
behavior and allows the player to get additional copies to put in the
shared stash.

It's worth noting this works per loot "component" and *not* by item ID.
So consider an enemy has the following loot table:

loot_count=2
loot=1,100
loot=1,100

The enemy will drop 2 of item ID 1, even if loot_drops_max=1 is in the
item definition.