Increase default audio playback samplerate to 44100Hz from 22050Hz
The reason for this change is that soundfx/no_mana.ogg stopped playing
for me. I don't know what about this particular file is causing trouble,
since we have other 44.1kHz files that down-sampled just fine. Opening
the SDL Mixer device as 44.1kHz seems to resolve all issues.

flare-game has a mix of 44100Hz and 22050Hz files, so conversion is
going to happen no matter what we choose. I guess it makes more sense to
increase the samplerate of the 22050Hz files than it does to decrease
the samplerate of the 44100Hz files.