PowerManager: Use std::vector instead of std::map for Powers
See #1871

The original goal of switching to map was to reduce memory usage when
there were large gaps in Power IDs. But element access with a map is
more expensive than a simple index into an "array" (a vector in our
case).

The reason memory usage was such an issue before was that our vector was
storing instances of the Power object. What this commit does is change
that to be Power pointers, only allocating the Powers that are defined.
This has a little memory overhead since we're storing a bunch of NULL
pointers for the undefined Powers, but it's not much (about 2kb for
flare-game).

To make sure we don't try to access the NULL entries, care has been
taken to use PowerManager::verifyID() everywhere that we read a power_id
from a config file (or save file).

For more memory savings, the Power class is now better packed.