Random item generation
Implements #1815

The generation of random items is done by starting with a source item
that has a `randomizer_def` defined. The randomizer then uses this file
to determine the level, quality, and bonuses that will be attached to
the item.

Here, we're introducing the concept of "extended items". This is
basically our way of storing items that have their properties modified
at runtime. Extended items are written whenever the game is saved in the
`extended_items.txt` file under the current game's save prefix. Random
items are the only current usage of this system, but the system should
support things such as sockets/identification/durability/etc.