Mouse movement re-work
Closes #1830
Closes flareteam/flare-game#789

Instead of having to hold the mouse button down, the player can now
click on a target and have them automatically pathfind their way to it.
In addition, if the player targets an event, a piece of loot, or an
enemy, they will perform the appropriate action when in range.

The behavior is meant to mimic Diablo II's behavior, as it is what many
new players may be familar with. I'll copy here the observations I made
in #1830:

- Don't need to hold the mouse button for the player to continue moving
- If the initial target **is not** an object that can be interacted
  with, holding the mouse button will continually update the pathfinding
  target
    - In this state, the player is locked to moving and can't interact
      with anything until the button is released
- If the initial target **is** an object that can be interacted with,
  holding the mouse button does nothing
    - In this state, the player is locked to moving towards the initial
      target and can't change trajectory until the button is released
    - The player will move until they are within range to perform the
      requested action. E.g. melee attacks and picking up loot require
      the player to move in close, but ranged attacks can be performed
      with little to no movement

Touchscreen users will also be able to take advantage of mouse-move now.
Even if touch controls are enabled in the settings, turning on
mouse-move will hide the virtual D-Pad and move around by touching where
they want to go. Since we can leave touch controls on in this scenario,
the player doesn't lose the touch behavior for UI (such as selecting
inventory items).