Consolidate checking power usage between Avatar and MenuActionBar
It makes more sense to check everything in MenuActionBar, because we
were already doing some checks here to determine if the slot should be
enabled or not. For the checks that don't affect the slots, we now do
those *before* the action is pushed to the action queue.

The handling of the action queue has also been simplified. Instead of
trying to have non-executed actions persist and updating their target,
we simply clear the action queue every frame. This is a lot easier to
reason about, and doesn't really change the end behavior.