Avatar: Pre-determine a list of valid powers for cooldown iteration
The IDs in the powers list might have large gaps depending on how the
modder defined their powers. The code touched in this commit gets called
every frame, so iterating over a list where most of the powers are
invalid is an expensive waste and can really slow things down. To
rememdy this, we can create a list of the valid IDs when we create the
timers and iterate over that instead.

The OpenValley mod highlights this well. We go from iterating over
600,000 items down to a little over 2,000!