Avatar: Process action queue before handling animation states
The main goal here is to be able to switch animation states in the same
frame that the action queue is processed. This removes the one frame
delay that would happen between a power being activated from the action
bar and the player entering the attack animation state.

This may need to be done for non-player entities as well, but I need to
test more.