Remove 1ms "compensation" from frame delay
Removing this make the SDL_Delay() call last 1ms longer, and reduces the
"rest of the frame" spinlock by 1ms. I don't think this should cause an
issue (if anything, it should become less likely that SDL_Delay() is
called with a value that the OS can't honor). The lower CPU usage under
normal circumstances is more desirable in my opinion.