Add itch.toml manifest (Windows only)
This file will be moved to ./.itch.toml for the Windows ZIPs as part of
the build process.

Closes #1857

I didn't add links to the mods folder or forums. The forums didn't seem
to open properly in the itch client. And I discovered that you can't use
enviroment variables for paths. So it was impossible to have a shortcut
to the *user's* mod folder as intended. It'd probably make more sense
anyways to have these shortcuts be in-game (maybe on the title screen?).