Animation: Refactor variable names and update some comments
Hopefully, this naming makes more sense. We were using "frame" to
describe both regular frames and, what we now call, sub-frames. The
former can be understood as the frames as defined in the animation data
files. Sub-frames, on the other hand, are generated by the engine and
advance once per render frame (which means it can vary based on max
FPS).

"Kinds" has also been renamed to "directions", as that is more in line
with how it is typically used. In fact, I made it a constant because we
always just passed 8. Of course, this may change if we decide to support
more (or less) directions.

Lastly, I rearranged the members for better memory packing. 16 bytes
smaller per Animation instance now.