Add auto-loot configuration option
Closes #1899

The option (located on the Game page of the config menu) has 3 options:
1. Disabled: no items will be auto-looted
2. Enabled: all eligible items will be auto-looted
 - Added auto_pickup property to items/item_types.txt to control this
3. Only currency: only currency will be auto-looted (i.e. the old
   behavior)

Note: the autopickup_currency property from engine/loot.txt still
applies. Since currency is not in items/types.txt, consider it the
equivalent of the auto_pickup property.

Note 2: Item type ID lookup is inefficient. This should be fixed!