GameState: Reset input key state when triggering a loading frame
Input doesn't get processed while we're in loading frames by design, so
the key state may be invalid by the time the loading is done (i.e. the
player lets go of a button before the loading is done). So we reset to
a known state where no keys are pressed or locked. This fixes the bug
where a given input (usually ACCEPT or CANCEL) needed to be pressed
twice to activate after having loaded into a new game state.