Animation: Use direction 0 for animations that aren't multi-direction
This addresses the issue where we want to be able to set a default NPC
direction, but flare-game's current NPC animations aren't
multi-directional. The result is that blank or garbage data gets
rendered when attempting to draw these NPCs at any direction other than
0.

The solution here is to use the data from direction 0 for the other
directions.