RenderDevice: Add support for multi-threaded image loading
For this initial commit, we are using it where it has the most impact,
the animation system. Results may vary, but I got about a 2.5x speedup
in opening GameStateLoad with several saves from flare-game with the new
HD art assets.

The usage is fairly simple: push images to load with
RenderDevice::pushQueuedImage() and then call
RenderDevice::loadQueuedImages() when ready to actually load them.

In case this causes problems, I've added an option to the user
settings.txt to disable it.