Revised control scheme for actionbar when using gamepad
Instead of a "modal" approach, we now constantly have the actionbar
tablist active while using a gamepad. Some of the button mappings have
changed:

- MENU_PAGE_PREV and MENU_PAGE_NEXT are used instead of the directional
  buttons
- ACTIONBAR is now used to open the menu for editing slots (e.g. move/clear)
- ACCEPT is only used when placing powers on the actionbar. It needs to
  be free in order to maintain its regular functionality.

A future commit will adjust these mappings so that they are in sensible
places on the gamepad itself.

This whole approach basically gives the player a customizable 3rd action
(aside from MAIN1/MAIN2) that they can easily trigger with
MENU_ACTIVATE. And more importantly, they can use it without having to
stop moving.

This commit also improves the tablist navigation a bit. Getting the
nearest widget now wraps around the screen. This is in contrast to the
old behavior that would select the next/prev widget in the list, which
may "skip" widgets that are closer to the screen edge.