RenderDevice: Don't call mods->locate() in loadQueuedImage() threads
mods->locate() mutates the state of ModManager, so we don't want to be
calling it in a thread. We should simple cache the result and use that
for the threads. Hopefully, this fixes the rare crash I've gotten
recently (didn't have debugger attached, but stack trace in system log
points towards this being the cause).