Improve parsing of data types with string-based IDs
There were two problems here:
1. It wasn't possible to modify an existing object via its ID. Instead,
   it would create a new object that may or may not have been used,
   depending on what order the list was processed later.
2. Even when not re-defining an object, the ID property needed to be
   first before any other data.

This commit fixes both. We now check for existing objects by ID, and
only create a new object if there's no match. And, as long as we're not
re-defining an existing object, the order of properites doesn't matter.
If the ID *does* match an existing object, we now log an error if we're
trying to parse non-ID data.