MenuActionBar: Try not to block MAIN1 and MAIN2
We don't want to stop the player from attacking if they are holding the
button and happen to accidentally place their cursor over the actionbar.
We should only lock the MAIN buttons when the player is not already
attacking *and* they click on one of the slots that can be interacted
with. For visual cleanliness, we can also hide actionbar tooltips while
the player is attacking, as well.

This commit also prevents the config menu button in the minimap menu
from being clicked if the player is attacking.