MenuActionBar: Fix 2-step power activation
Not sure when this broke, but MAIN1 wasn't activating the second stage
as it should. This commit makes it so that both MAIN1 and MAIN2 can
perform the second stage. It makes sense to do this since the first
stage uses MAIN2, so the player might try to use the same button.