MapRenderer: Replace entity markers with wall fading for player
Closes #1909

Note that this only applies to the player. I think it makes more sense
that the player can reveal enemies that are behind walls by being in
close proximity to them. Plus, the player can always look at their
minimap.

The minimum opacity can be controlled with the 'fade_wall_alpha'
property in engine/misc.txt. There is also a video option for the
player if they wish to turn this feature off.