InputState: Add 'cycle_menus' key binding
This bind allows players to use a single button to access all of the
primary menus. If no menus are open, the Character and Inventory menus
will open (if they're enabled). On the second press, the Log and Powers
menus will open. On the third press (or if any other combination of
windows is open), all windows will be closed.

The primary use for this will be for gamepads, since we're strapped for
available buttons there.