StatBlock: Add fallback method for hero base/head graphics
Fixes #1908

If there are no animations found for a given base_gfx value, we try to
fall back to either "female" or "male". "female" will be used if the
base_gfx starts with that string, otherwise "male" will be used.

If the base_head can't be found, we simply try to fall back to
"head_short".

TODO:
- We should probably determine the fallback values based on the actual
  mod contents (i.e. if there's no "male" or "head_short").
- Maybe this function should be used to ensure that base graphics match
  a mod's engine/hero_options.txt contents?