ItemManager: Fix extended items from other saves being deleted
Fixes #1913

This works by setting a flag on all items that represents whether or not
they are "foreign". We assume all items are foreign by default, and
*only* unset this flag if:

- the item was initially loaded into the inventory or stash
- it was generated during the current game session

When saving the extended items file, we now only skip writing the item
data if the item is not foreign. Otherwise, it gets preserved.

There is still the issue where deleting a save does not clean up the
extended items from that save, but that is a separate problem.